canvas生成を削除して修正
