+++ src/components/features/play/StartCountdown/index.tsx
@@ -3,7 +3,9 @@ import styles from "./index.module.css";
 export const StartCountdown = ({ startTime }: { startTime: number }) => {
 	return (
 		<div className={styles.root}>
-			<p>{startTime}</p>
+			<p className={styles.text} data-time={startTime}>
+				{startTime !== 0 ? startTime : "スタート"}
+			</p>
 		</div>
 	);
 };
