+++ src/pages/PlayTimeAttack/index.tsx
@@ -1,32 +1,27 @@
-import { useEffect } from "react";
-import { useTimeAttackStore } from "../../store/useTimeAttackStore";
 import { Timer } from "../../components/features/play/Timer";
 import { ARfunction } from "../AR";
-import styles from "./index.module.css";
 import ExitButton from "../../components/features/play/ExitButton";
-import { useUserStore } from "../../store/useUserStore";
-import { useSocketRefStore } from "../../store/useSocketRefStore";
+import { StartCountdown } from "../../components/features/play/StartCountdown";
+import { usePlayTimeAttack } from "./hooks";
+import styles from "./index.module.css";
+
+const gameDuration = 11;
 
 export const PlayTimeAttack = () => {
-  const setClearTime = useTimeAttackStore((state) => state.setClearTime);
-  const { name: userName } = useUserStore();
-  const time = useSocketRefStore((state) => state.eventState.time);
+	const { time, userName } = usePlayTimeAttack();
 
-  // timeが0になったらrankingページへ遷移する
-  useEffect(() => {
-    setClearTime(18);
-
-    // if (time === 0) {
-    //   navigate("/ranking_timeAttack");
-    // }
-  }, [setClearTime]);
-
-  return (
-    <main>
-      <Timer remainingTime={time} />
-      <p className={styles["user-name"]}>{userName}</p>
-      <ARfunction />
-      <ExitButton />
-    </main>
-  );
+	return (
+		<main>
+			{time <= 0 ? (
+				<StartCountdown startTime={-time} />
+			) : (
+				<>
+					<Timer remainingTime={gameDuration - time} />
+					<p className={styles["user-name"]}>{userName}</p>
+					<ARfunction />
+					<ExitButton />
+				</>
+			)}
+		</main>
+	);
 };
