+++ src/proto/game/game_pb.d.ts
@@ -1,6 +1,5 @@
 // @generated by protoc-gen-es v2.0.0
 // @generated from file game/game.proto (package game, syntax proto3)
- 
 
 import type { GenFile, GenService } from "@bufbuild/protobuf/codegenv1";
 import type {
