+++ src/proto/game/resources/game_pb.d.ts
@@ -1,6 +1,5 @@
 // @generated by protoc-gen-es v2.0.0
 // @generated from file game/resources/game.proto (package game.resources, syntax proto3)
- 
 
 import type {
 	GenEnum,
