+++ src/proto/game/rpc/game_pb.d.ts
@@ -1,6 +1,5 @@
 // @generated by protoc-gen-es v2.0.0
 // @generated from file game/rpc/game.proto (package game.rpc, syntax proto3)
- 
 
 import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
 import type { Message } from "@bufbuild/protobuf";
